+++ packages/antd-management-fast-design-playground/CHANGELOG.md
@@ -1,5 +1,16 @@
 # antd-management-fast-design-playground
 
+## 1.0.114
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-react@1.0.72
+  - antd-management-fast-component@2.6.161
+  - antd-management-fast-formily@1.0.74
+  - antd-management-fast-common@2.6.104
+
 ## 1.0.113
 
 ### Patch Changes